player: fix seeking when "cut" is enabled
when auto cutting silence was enabled, tracks seemd to start at random
positins or were skipped completely. Actually they started at the offset
where the previous one was cut. This was due to the Gstreamer pipeline
remembering the previous position. When starting a new file, it tried to
resume playback from exactly this position, skipping the track completely
when the position exceeded it's duration.

Furthermote seg_to was ignored when seg_from wasn't set.